AAS_PointInsideFace: cast loop counter compare to int (jl, not jb)
v5 is _DWORD * (unsigned), so v5[2] is unsigned and the C promotion
rules force the compare to unsigned (jb), but the original code emits
jl (signed). Casting v5[2] to int restores the signed compare,
matching the original byte sequence.

-1 byte_diff (163902 → 163901).

Source-commit: 83b4e0bc7c69b892ecde9483e4b695afd09b1b47

+++ botlib/botlib.c
@@ -17571,7 +17571,7 @@ qboolean __cdecl AAS_PointInsideFace(int facenum, vec3_t point, float epsilon)
       v17 = -epsilon;
       if ( v15[2] * v13 + v15[1] * v12 + v15[0] * v11 < v17 )
         break;
-      if ( ++v16 >= v5[2] )
+      if ( ++v16 >= (int)v5[2] )
         return 1;
     }
     return 0;
